feat(expr): builtins and implicit arg
If a function takes a single argument and the `self` psuedo-value is
present in the scope, an attempt is made to use that as the missing
argument. This is to support use cases in env where we might not want to
implicitly pass an argument since the presence of `self` indicates it is
running in something like an env struct field.

Also adds a lot of builtins. The expr "standard library" is still a mess
but I'm going to keep refining what is there. In addition to usability,
more consideration will need to be given to security and testing as
well.